requirements.txt
dotenv
openai-agents
tree-sitter==0.20.4
pyright